pyproject.toml
[tool.poetry]
name = "coc-discord-bot"
version = "0.1.0"
description = ""
authors = ["Rizwan Riaz <iamrizwan077@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.1.1"
requests = "^2.32.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
